SCRUM-35 Work in Progress: Adjusting the file structure to become more scalable and maintainable as our models continue to grow. Going to do the same thing with the views. I am leaving all current files as well, to ensure that the migrations and seeding still working properly. Going to adjust the migrations with all files included before deleting old files.
