+++ backend/app/models/customer_models.py
@@ -1,0 +1,28 @@
+# backend/app/models.py
+
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+class CustomUser(AbstractUser):
+    phone = models.CharField(max_length=20, blank=True, null=True)
+    business_name = models.CharField(max_length=255, blank=True, null=True)
+    business_address = models.TextField(blank=True, null=True)
+
+
+class Customer(models.Model):
+    user = models.OneToOneField(CustomUser, on_delete=models.CASCADE, related_name="customer")
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    def to_dict(self):
+        return {
+            "id": self.id,
+            "username": self.user.username,
+            "email": self.user.email,
+            "phone": self.user.phone,
+            "created_at": self.created_at,
+            "updated_at": self.updated_at,
+        }
+
+    def __str__(self):
+        return f"{self.user.username}'s Customer Profile"
