+++ backend/app/models/restaurant_models.py
@@ -1,31 +1,5 @@
-# backend/app/models.py
-
-from django.contrib.auth.models import AbstractUser
 from django.db import models
-
-class CustomUser(AbstractUser):
-    phone = models.CharField(max_length=20, blank=True, null=True)
-    business_name = models.CharField(max_length=255, blank=True, null=True)
-    business_address = models.TextField(blank=True, null=True)
-
-
-class Customer(models.Model):
-    user = models.OneToOneField(CustomUser, on_delete=models.CASCADE, related_name="customer")
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def to_dict(self):
-        return {
-            "id": self.id,
-            "username": self.user.username,
-            "email": self.user.email,
-            "phone": self.user.phone,
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-        }
-
-    def __str__(self):
-        return f"{self.user.username}'s Customer Profile"
+from .customer_models import Customer
 
 class Restaurant(models.Model):
     name = models.CharField(max_length=255)
@@ -77,5 +51,11 @@ class OrderItems(models.Model):
 
     def __str__(self):
         return f"{self.quantity} of {self.item.name} from {self.restaurant.name}"
-
-
+    
+class CurrentItem(models.Model):
+    item_name = models.CharField(max_length=100)
+    quantity = models.PositiveIntegerField(default=1)
+    status = models.CharField(max_length=20, default="active")
+    created_at = models.DateTimeField(auto_now_add=True)
+    def __str__(self):
+        return f"{self.item_name} ({self.quantity})"
